fix: status and cid are strings in the query result (#21)
The types expect `status` and `cid` to be arrays of strings, but they
arrive as strings so the code fails.

Instead of using multiple query args we use a single are with multiple
values delimited by commas.  Problem is this seems to be broken in
oas-tools at the moment: https://github.com/oas-tools/oas-tools/issues/248

This PR works around the breakage.